+++ avs_core/filters/conditional/conditional_functions.cpp
@@ -405,11 +405,11 @@ AVSValue ComparePlane::CmpPlaneSame(AVSValue clip, void* user_data, int offset, 
 // Y Planes functions
 
 AVSValue __cdecl MinMaxPlane::Create_max_y(AVSValue args, void* user_data, IScriptEnvironment* env) {
-  return MinMax(args[0],user_data, args[1].AsFloat(0.0f), PLANAR_Y, MAX, env);
+  return MinMax(args[0],user_data, (float)args[1].AsFloat(0.0f), PLANAR_Y, MAX, env);
 }
 
 AVSValue __cdecl MinMaxPlane::Create_min_y(AVSValue args, void* user_data, IScriptEnvironment* env) {
-  return MinMax(args[0],user_data, args[1].AsFloat(0.0f), PLANAR_Y, MIN, env);
+  return MinMax(args[0],user_data, (float)args[1].AsFloat(0.0f), PLANAR_Y, MIN, env);
 }
 
 AVSValue __cdecl MinMaxPlane::Create_median_y(AVSValue args, void* user_data, IScriptEnvironment* env) {
@@ -417,17 +417,17 @@ AVSValue __cdecl MinMaxPlane::Create_median_y(AVSValue args, void* user_data, IS
 }
 
 AVSValue __cdecl MinMaxPlane::Create_minmax_y(AVSValue args, void* user_data, IScriptEnvironment* env) {
-  return MinMax(args[0],user_data, args[1].AsFloat(0.0f), PLANAR_Y, MINMAX_DIFFERENCE, env);
+  return MinMax(args[0],user_data, (float)args[1].AsFloat(0.0f), PLANAR_Y, MINMAX_DIFFERENCE, env);
 }
 
 // U Planes functions
 
 AVSValue __cdecl MinMaxPlane::Create_max_u(AVSValue args, void* user_data, IScriptEnvironment* env) {
-  return MinMax(args[0],user_data, args[1].AsFloat(0.0f), PLANAR_U, MAX, env);
+  return MinMax(args[0],user_data, (float)args[1].AsFloat(0.0f), PLANAR_U, MAX, env);
 }
 
 AVSValue __cdecl MinMaxPlane::Create_min_u(AVSValue args, void* user_data, IScriptEnvironment* env) {
-  return MinMax(args[0],user_data, args[1].AsFloat(0.0f), PLANAR_U, MIN, env);
+  return MinMax(args[0],user_data, (float)args[1].AsFloat(0.0f), PLANAR_U, MIN, env);
 }
 
 AVSValue __cdecl MinMaxPlane::Create_median_u(AVSValue args, void* user_data, IScriptEnvironment* env) {
@@ -435,16 +435,16 @@ AVSValue __cdecl MinMaxPlane::Create_median_u(AVSValue args, void* user_data, IS
 }
 
 AVSValue __cdecl MinMaxPlane::Create_minmax_u(AVSValue args, void* user_data, IScriptEnvironment* env) {
-  return MinMax(args[0],user_data, args[1].AsFloat(0.0f), PLANAR_U, MINMAX_DIFFERENCE, env);
+  return MinMax(args[0],user_data, (float)args[1].AsFloat(0.0f), PLANAR_U, MINMAX_DIFFERENCE, env);
 }
 // V Planes functions
 
 AVSValue __cdecl MinMaxPlane::Create_max_v(AVSValue args, void* user_data, IScriptEnvironment* env) {
-  return MinMax(args[0],user_data, args[1].AsFloat(0.0f), PLANAR_V, MAX, env);
+  return MinMax(args[0],user_data, (float)args[1].AsFloat(0.0f), PLANAR_V, MAX, env);
 }
 
 AVSValue __cdecl MinMaxPlane::Create_min_v(AVSValue args, void* user_data, IScriptEnvironment* env) {
-  return MinMax(args[0],user_data, args[1].AsFloat(0.0f), PLANAR_V, MIN, env);
+  return MinMax(args[0],user_data, (float)args[1].AsFloat(0.0f), PLANAR_V, MIN, env);
 }
 
 AVSValue __cdecl MinMaxPlane::Create_median_v(AVSValue args, void* user_data, IScriptEnvironment* env) {
@@ -452,7 +452,7 @@ AVSValue __cdecl MinMaxPlane::Create_median_v(AVSValue args, void* user_data, IS
 }
 
 AVSValue __cdecl MinMaxPlane::Create_minmax_v(AVSValue args, void* user_data, IScriptEnvironment* env) {
-  return MinMax(args[0],user_data, args[1].AsFloat(0.0f), PLANAR_V, MINMAX_DIFFERENCE, env);
+  return MinMax(args[0],user_data, (float)args[1].AsFloat(0.0f), PLANAR_V, MINMAX_DIFFERENCE, env);
 }
