+++ avs_core/filters/debug.cpp
@@ -187,7 +187,7 @@ public:
   }
 
   static AVSValue __cdecl Create(AVSValue args, void*, IScriptEnvironment* env) {
-    return new Preroll(args[0].AsClip(), args[1].AsInt(0), args[2].AsDblDef(0.0), env);
+    return new Preroll(args[0].AsClip(), args[1].AsInt(0), (float)args[2].AsDblDef(0.0), env);
   }
 
 };
